+++ CMakeLists.txt
@@ -257,6 +257,8 @@ set(WALLET_SOURCES
         ./src/wallet/db.cpp
         ./src/addressbook.cpp
         ./src/crypter.cpp
+        ./src/budget/budgetvote.cpp
+        ./src/budget/finalizedbudgetvote.cpp
         ./src/masternode.cpp
         ./src/masternode-budget.cpp
         ./src/masternode-payments.cpp
